Use procs to decorate GPX result from GeoJSON
GeoJSON often has ad hoc data stored in a "properties" object within each feature. Some of this data can be useful when transferred to a GPX file. Since there is no universal way to transfer the data, this commit allows a developer to specify an object that responds to "call" (e.g., lambdas or procs) to decorate a GPX feature during the conversion. The following are available:

- line_string_feature_to_segment
- multi_line_string_feature_to_track
- point_feature_to_waypoint
- multi_point_feature_to_waypoint